Revert "Fixes #182"
This reverts commit 02041571cec76696adf30c3e5f55cdb13e144d20.

+++ src/closure.h
@@ -16,7 +16,7 @@ typedef struct {
   Persistent<Function> pfn;
   Handle<Function> fn;
   unsigned len;
-  int16_t max_len;
+  unsigned max_len;
   uint8_t *data;
   Canvas *canvas;
   cairo_status_t status;
